requirements.txt
anyio==4.6.2.post1
attrs==24.2.0
blinker==1.9.0
certifi==2024.8.30
chardet==4.0.0
charset-normalizer==3.4.0
click==8.1.7
dataclasses==0.6
deprecation==2.1.0
Flask==3.1.0
gotrue==0.2.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
idna==2.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
packaging==24.2
pluggy==1.5.0
postgrest-py==0.4.0
psycopg2-binary==2.9.10
py==1.11.0
pydantic==1.10.19
pytest==6.2.5
python-dateutil==2.9.0.post0
python-telegram-bot==21.7
realtime-py==0.1.3
requests==2.25.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.1
StrEnum==0.4.15
supabase-py==0.0.2
toml==0.10.2
typing_extensions==4.12.2
urllib3==1.26.20
websockets==9.1
Werkzeug==3.1.3
